Move showModalBottomSheet into separate file
