+++ lib/features/app/presentation/widgets/misc/AppBar/settings_button.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:weathque/config/theme/custom_colors.dart';
+import 'package:weathque/features/app/presentation/widgets/misc/bottom_sheet/bottom_sheet.dart';
 
 class SettingsButton extends StatefulWidget {
   const SettingsButton({super.key});
@@ -45,56 +45,11 @@ class _SettingsButtonState extends State<SettingsButton> with TickerProviderStat
           _controller.reset();
           _controller.forward();
 
-          showModalBottomSheet(
-            context: context, 
-            isScrollControlled: true,
-            shape: RoundedRectangleBorder(
-              borderRadius: BorderRadius.vertical(top: Radius.circular(20.0)),
-            ),
-            backgroundColor: CustomColors.yellow.color,
-            builder:(context) {
-              return _buildBottomSheetMenu(context);
-            },
-          );
+          showCustomBottomSheet(context);
         },
       ),
     );
   }
 
-  Widget _buildBottomSheetMenu(BuildContext context){
-    return Container(
-      width: double.infinity,
-      height: MediaQuery.of(context).size.height - (kBottomNavigationBarHeight + kToolbarHeight),
-      child: Padding(
-        padding: EdgeInsets.symmetric(vertical: 30, horizontal: 15),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.center,
-          children: [
-            TextField(
-              cursorColor: Colors.black,
-              decoration: InputDecoration(
-                enabledBorder: OutlineInputBorder(
-                  borderRadius: BorderRadius.all(Radius.circular(20.0)),
-                  borderSide: BorderSide(color: Colors.black)
-                ),
-                focusedBorder: OutlineInputBorder(
-                  borderRadius: BorderRadius.all(Radius.circular(20.0)),
-                  borderSide: BorderSide(color: Colors.black)
-                ),
-                hintText: "Enter a city",
-                labelText: "City",
-                labelStyle: TextStyle(
-                  color: Colors.black
-                ),
-                suffixIcon: IconButton(
-                  icon: Icon(Icons.search, color: Colors.black),
-                  onPressed: (){},
-                )
-              ),
-            )
-          ],
-        ),
-      ),
-    );
-  }
+  
 }
